Updated LHA Reader
