+++ DataBase/lha_blocks.cpp
@@ -1,12 +1,8 @@
-#include <iostream>
-#include <iomanip>
-#include <fstream>
 #include <string>
 #include <sstream>
 #include <memory>
 #include <algorithm>
-#include <span>
-#include <complex>
+
 #include "lha_blocks.h"
 #include "lha_elements.h"
